small refactor to move the transport resource into the transport.rs file
